+++ compliance/runtests.py
@@ -7,6 +7,9 @@ import sys
 
 from six.moves.urllib_parse import urlencode
 
+logging.basicConfig()
+logging.getLogger('lomond').setLevel(logging.INFO)
+
 from lomond import WebSocket
 from lomond.constants import USER_AGENT
 
@@ -21,6 +24,7 @@ def get_test_count():
     ws = WebSocket(server + '/getCaseCount')
     case_count = None
     for event in ws:
+        print(event)
         if event.name == 'text':
             case_count = json.loads(event.text)
     if case_count is None:
@@ -42,13 +46,13 @@ def run_tests():
 
 def run_ws(url):
     """Run a websocket until close."""
-    ws = WebSocket(url)
+    ws = WebSocket(url, compress=True)
     for event in ws.connect(ping_rate=0):
         try:
             if event.name == 'text':
-                ws.send_text(event.text)
+                ws.send_text(event.text, compress=False)
             elif event.name == 'binary':
-                ws.send_binary(event.data)
+                ws.send_binary(event.data, compress=False)
         except:
             log.exception('error running websocket')
             break