+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.0] - Unreleased
+
+### Added
+
+- Websocket compression
+
+### Changed
+
+- Response now uses unicode rather thatn bytes for keys/value
+
 ## [0.2.5] - 2018-05-28
 
 ### Fixed